+++ src/commonMain/kotlin/csense/kotlin/extensions/collections/array/FloatArray.kt
@@ -15,22 +15,6 @@ public inline fun <U> FloatArray.forEachDiscard(receiver: Function1<Float, U>): 
 
 //region Generic collection extensions
 /**
- * Performs traversal in pairs of 2  (with the first index as well)
- */
-@Deprecated("will be removed")
-@Suppress("MissingTestFunction")
-public inline fun FloatArray.forEach2Indexed(action: Function2IndexedUnit<Float, Float>): Unit =
-    GenericCollectionExtensions.forEach2Indexed(count(), ::elementAt, action)
-
-/**
- * Performs traversal in pairs of 2
- */
-@Deprecated("will be removed")
-@Suppress("MissingTestFunction")
-public inline fun FloatArray.forEach2(action: Function2Unit<Float, Float>): Unit =
-    GenericCollectionExtensions.forEach2(count(), ::elementAt, action)
-
-/**
  * Performs backwards traversal on this [FloatArray].
  */
 public inline fun FloatArray.forEachBackwards(action: FunctionUnit<Float>): Unit =