+++ src/commonMain/kotlin/csense/kotlin/extensions/coroutines/Coroutines.kt
@@ -54,9 +54,11 @@ public suspend fun <T, U> Iterable<T>.mapAsyncAwait(
 
 /**
  * Waits for all jobs in an array.
- * Since this is missing from Standard Library.
+ * Since this is missing from Standard Library...
  * @receiver [Array]<out [Job]>
  */
 public suspend fun Array<out Job>.joinAll() {
     forEach { it.join() }
 }
+
+